requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit-lottie